requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
couchbase==4.1.7
python-dotenv==1.0.0
uuid==1.30